+++ CHANGELOG.md
@@ -5,6 +5,7 @@ omnibus-gitlab repository.
 
 8.10.0
 
+- Fix RangeError bignum too big errors on armhf platforms 4ba24bfe
 - Update redis to 3.2.1 (Takuya Noguchi)
 - Updated Chef version to 12.10.24 6e0c66
 - Disable nodejs Snapshot feature on ARM platforms f9a7b4bf
@@ -39,6 +40,7 @@ omnibus-gitlab repository.
 
 - Make default IMAP incoming mailbox "inbox" in case user omits this setting d3c187
 - Make NGINX server_names_hash_bucket_size configurable and default it to 64 bytes 7cb488
+- Use gitlab:db:configure to seed and migrate the database
 - Add log prefix for pages and registry services 48e29b
 - Add configuration option for the Container Registry storage driver
 - Change the autovacuum configuration defaults f5ac85